Refactor Track data class and TrackButton to use Double for track size; update related components for improved state management

+++ shared/src/main/java/com/ontrek/shared/data/Track.kt
@@ -14,6 +14,6 @@ data class Track(
     val filename: String,
     val stats: TrackStats,
     val title: String,
-    val size: Long,
+    val size: Double,
     val upload_date: String
 )
